Rewrite `instanceof` and add unchecked casting
`instanceof` was rewritten so that it may be called from all contexts of `Invocation`, while unchecked casting was added to `Unchecked` instead. These changes allowed the javadoc that annotates each part of an invocation to become clearer and more specific.

This commit also adds the `<dependencies>` section, which lets you specify additional dependencies to scan, and the `removals` category, which usages of `Label` and `Unchecked` are counted for.